jitter for LR

+++ notebooks/homhots_LR_jitter.py
@@ -1,8 +1,9 @@
 import numpy as np
 import sys
 sys.path.append('../HOTS')
-from Tools import tic,toc, get_loader, fit_data, predict_data, classification_results, netparam
+from Tools import tic, toc, get_loader, fit_data, predict_data, classification_results, netparam
 import pickle
+from os.path import isfile
 
 if __name__ == '__main__':
     #_________NETWORK_PARAMETERS______________________
@@ -17,12 +18,6 @@ if __name__ == '__main__':
     R = 2
     nbclust = [4, 8, 16]
     filt = 2
-    
-    #_______________JITTER_________________________
-    jit_s = np.arange(0,10,0.5)
-    jit_t = np.arange(0,300,10)
-    jit_s, jit_t = jit_s**2, jit_t**2
-    #______________________________________________
 
     #_______________NB_OF_DIGITS___________________
     dataset = 'nmnist'
@@ -32,6 +27,8 @@ if __name__ == '__main__':
     nb_test = nb_test//ds
     nb_train = nb_train//ds
     print(f'training set size: {nb_train} - testing set: {nb_test}')
+    subset_size = ds
+    nb_trials = ds
     #______________________________________________
     #_______________LR_PARAMETERS__________________
     num_workers = 0
@@ -42,58 +39,48 @@ if __name__ == '__main__':
     #num_epochs = 2 ** 9 + 1
     print(f'number of epochs: {num_epochs}')
     #______________________________________________
-    
+
+
     timestr = '2021-03-29'
-    ds_ev_output = 1
-    record_path = '../Records/EXP_03_NMNIST/models/'
-    
-    nb_trials = 10
-    results_s = np.zeros([nb_trials, len(jit_s)])
-    results_t = np.zeros([nb_trials, len(jit_t)])
-    
-
-    for name in ['homhots']:
-        #learn_set, nb_pola, name_net = get_loader(name, record_path, nb_train, True, filt, tau, nbclust, sigma, homeinv, jitter, timestr, dataset, R, ds_ev = ds_ev_output)
-        param = [0.25, 1]
-        tau_nm = [5.0, 50.0, 500.0]
-        R_nm = [2, 4, 8]
-        name_net = f'../Records/EXP_03_NMNIST/models/2021-03-29_lagorce_rdn_None_True_{param}_{nbclust}_{tau_nm}_{R_nm}_False_LR_60000.pkl'
-        with open(name_net, 'rb') as file:
-            model, loss = pickle.load(file)
-        
-        #model, loss = fit_data(name_net, learn_set, nb_train, nb_pola, learning_rate, num_epochs, betas, num_workers=num_workers, verbose=True)
-
-        for trial in [0]:
-            timestr = '2021-03-29_'+str(trial)
-            id_jit = 0
-            for i in jit_s:
-                i = round(i,2)
-                jitonic = [None,i]
-                if i==0:
-                    jitonic = [None,None]
-                test_set, nb_pola, name_net = get_loader(name, record_path, nb_test, False, filt, tau, nbclust, sigma, homeinv, jitter, timestr, dataset, R, subset_size = nb_test, jitonic = jitonic, ds_ev = ds_ev_output)
-                print(f'prediction for {name}...')
-                pred_target, true_target = predict_data(test_set, model, nb_test, num_workers=num_workers)
-                meanac, onlinac, lastac = classification_results(pred_target, true_target, nb_test)
-                print(f'Classification performance for {name}: {meanac}')
-                results_s[trial,id_jit] = meanac
-                id_jit += 1
-                
-            id_jit = 0
-            for j in jit_t:
-                j = round(j,0)
-                jitonic = [j,None]
-                if j==0:
-                    jitonic = [None,None]
-                test_set, nb_pola, name_net = get_loader(name, record_path, nb_test, False, filt, tau, nbclust, sigma, homeinv, jitter, timestr, dataset, R, subset_size = nb_test, jitonic = jitonic, ds_ev = ds_ev_output)
-                print(f'prediction for {name}...')
-                pred_target, true_target = predict_data(test_set, model, nb_test, num_workers=num_workers)
-                meanac, onlinac, lastac = classification_results(pred_target, true_target, nb_test)
-                print(f'Classification performance for {name}: {meanac}')
-                results_t[trial,id_jit] = meanac
-
     path = '../Records/EXP_03_NMNIST/'
-    f_name = f'{path}{timestr}_LR_results_jitter_{nbclust}_{nb_train}_{nb_test}_{ds_ev_output}.pkl'
+    
+    thres = None
 
-    with open(f_name, 'wb') as file:
-        pickle.dump([results_s, results_t], file, pickle.HIGHEST_PROTOCOL)
+    for name in ['homhots','hots','raw']:
+        f_name = f'{path}{timestr}_LR_results_jitter_{name}_{nbclust}_{nb_train}_{nb_test}_{ds_ev}_{thres}.pkl'
+        if isfile(f_name):
+            with open(f_name, 'rb') as file:
+                likelihood, true_target = pickle.load(file)
+        else:
+            ds_ev = 10
+            print(f'LR fit for {name}...')
+            model, loss  = fit_data(name,timestr,path,filt,tau,R,nbclust,sigma,homeinv,jitter,dataset,nb_train, ds_ev,learning_rate,num_epochs,betas, tau_cla,jitonic=jitonic,subset_size=nb_train,num_workers=num_workers,verbose=False)
+            
+            ds_ev = 1
+            results_s, results_s_last = np.zeros([2, nb_trials, len(jit_s)])
+            results_t, results_t_last = np.zeros([2, nb_trials, len(jit_t)])
+            for trial in range(nb_trials):
+                timestr = '2021-03-29_'+str(trial)
+                id_jit = 0
+                for id_jit, i in enumerate(jit_s):
+                    i = round(i,2)
+                    jitonic = [None,i]
+                    if i==0:
+                        jitonic = [None,None]
+                    likelihood, true_target = predict_data(model,name,timestr,path,filt,tau,R,nbclust,sigma, homeinv, jitter,dataset,nb_test,ds_ev,tau_cla,jitonic=jitonic,subset_size=nb_test,num_workers=num_workers, verbose=False)
+                    meanac, onlinac, lastac, truepos, falsepos = classification_results(likelihood, true_target, thres, nb_test)
+                    results_s[trial,id_jit] = meanac
+                    results_s_last[trial,id_jit] = lastac
+
+                for id_jit, i in enumerate(jit_t):
+                    j = round(j,0)
+                    jitonic = [j,None]
+                    if j==0:
+                        jitonic = [None,None]
+                    likelihood, true_target = predict_data(model,name,timestr,path,filt,tau,R,nbclust,sigma, homeinv, jitter,dataset,nb_test,ds_ev,tau_cla,jitonic=jitonic,subset_size=nb_test,num_workers=num_workers, verbose=False)
+                    meanac, onlinac, lastac, truepos, falsepos = classification_results(likelihood, true_target, thres, nb_test)
+                    results_t[trial,id_jit] = meanac
+                    results_t_last[trial,id_jit] = lastac
+
+            with open(f_name, 'wb') as file:
+                pickle.dump([results_s, results_t, results_s_last, results_t_last], file, pickle.HIGHEST_PROTOCOL)
